statedb: Move benchmarks to benchmarks_test.go
The db_test.go is starting to get a bit too big.
